+++ src/components/auth/authFormItem/authFormItem.styled.ts
@@ -1,0 +1,24 @@
+import { css } from 'styled-components';
+
+const label = css`
+  width: 100%;
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+`;
+
+const wrapper = css`
+  width: 660px;
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+`;
+
+const item = css`
+  width: 100%;
+  box-sizing: border-box;
+  display: flex;
+  gap: 20px;
+`;
+
+export { item, label, wrapper };
