+++ src/components/auth/authFormItem/index.ts
@@ -1,0 +1,1 @@
+export {};
